Use IVsSolution to find the project being profiled, so that we consider projects located under solution folders. This matches the code that is used to populate the project list in the start profiling dialog.
